% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_propulmeta.R
\docType{data}
\name{amlodipine}
\alias{amlodipine}
\title{Amlodipine Clinical Trial Dataset}
\format{
A data frame with 8 rows and 7 variables:
\describe{
\item{study}{Study label}
\item{n.amlo}{Sample size in Amlodipine group}
\item{mean.amlo}{Mean change in Amlodipine group}
\item{var.amlo}{Variance in Amlodipine group}
\item{n.plac}{Sample size in placebo group}
\item{mean.plac}{Mean change in placebo group}
\item{var.plac}{Variance in placebo group}
}
}
\source{
Normand, S.L.T. (1999). Meta-analysis of Amlodipine trials.
}
\usage{
data(amlodipine)
}
\description{
Data from clinical trials comparing Amlodipine to placebo.
}
\keyword{datasets}
