% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{bubble_plot}
\alias{bubble_plot}
\title{Bubble Plot for Meta-Regression}
\usage{
bubble_plot(
  meta_reg_object,
  moderator = NULL,
  plot_all_levels = TRUE,
  save_as = c("viewer", "pdf", "png"),
  filename = NULL,
  width = 10,
  height = 8,
  ...
)
}
\arguments{
\item{meta_reg_object}{A \code{meta_reg} object created with \code{meta_reg()}.}

\item{moderator}{Character. Name of moderator variable (required if multiple predictors).}

\item{plot_all_levels}{Logical. For categorical moderators, plot each dummy variable (default = TRUE).}

\item{save_as}{"viewer", "pdf", or "png". Default is "viewer".}

\item{filename}{Optional file name if exporting.}

\item{width, height}{Dimensions in inches for export. Defaults: width = 10, height = 8.}

\item{...}{Additional arguments passed to \code{metafor::regplot()}.}
}
\value{
A bubble plot rendered or saved.
}
\description{
Displays a bubble plot for a \code{meta_reg} object using \code{metafor::regplot()}.
}
