% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_propulmeta.R
\docType{data}
\name{caffeine}
\alias{caffeine}
\title{Caffeine and Endurance Performance Dataset}
\format{
An object of class \code{data.frame} with 8 rows and 12 columns.
}
\source{
Adapted from \code{meta} package example datasets.
}
\usage{
data(caffeine)
}
\description{
This dataset contains data from randomized trials evaluating the effect of caffeine
ingestion on endurance performance, measured as time to exhaustion or distance covered.
}
\details{
#' @format A data frame with 8 rows and 12 variables:
\describe{
\item{study}{Study label}
\item{year}{Year of publication}
\item{h.caf}{Number of events in the caffeine group}
\item{n.caf}{Sample size in the caffeine group}
\item{h.decaf}{Number of events in the decaffeinated group}
\item{n.decaf}{Sample size in the decaffeinated group}
\item{D1}{Risk of bias domain 1 (e.g., "low", "some", "high")}
\item{D2}{Risk of bias domain 2}
\item{D3}{Risk of bias domain 3}
\item{D4}{Risk of bias domain 4}
\item{D5}{Risk of bias domain 5}
\item{rob}{Overall risk of bias}
}

This dataset is commonly used in meta-analyses of continuous outcomes to demonstrate methods for pooling mean differences
or standardized mean differences.
}
\examples{
data(caffeine)
head(caffeine)
}
\keyword{datasets}
