% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_propulmeta.R
\docType{data}
\name{dat_bcg}
\alias{dat_bcg}
\title{Example Dataset: dat_bcg}
\format{
A data frame with 13 rows and 11 variables:
\describe{
\item{trial}{Trial ID}
\item{author}{First author of the study}
\item{year}{Year of publication}
\item{tpos}{Number of tuberculosis cases in the treatment group (BCG vaccinated)}
\item{tneg}{Number without tuberculosis in the treatment group}
\item{cpos}{Number of tuberculosis cases in the control group (not vaccinated)}
\item{cneg}{Number without tuberculosis in the control group}
\item{ablat}{Latitude of the study location}
\item{alloc}{Randomization method (e.g., random, alternate)}
\item{npos}{Total sample size in the treatment group}
\item{region}{Region of the study (e.g., North America, Europe)}
}
}
\source{
Colditz et al., 1994. Meta-analysis of BCG vaccine efficacy.
}
\usage{
data(dat_bcg)
}
\description{
This dataset contains results from clinical trials evaluating the efficacy of Bacillus Calmette-Guérin (BCG) vaccine against tuberculosis.
}
\details{
This dataset is commonly used in meta-analysis examples to explore heterogeneity of BCG vaccine efficacy across different regions.
}
\keyword{datasets}
