% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_propulmeta.R
\docType{data}
\name{dat_normand1999}
\alias{dat_normand1999}
\title{Normand 1999 Meta-analysis Dataset (Continuous Outcomes)}
\format{
A data frame with 9 rows and 8 variables:
\describe{
\item{study}{Study ID}
\item{source}{Study source or name}
\item{n1i}{Sample size in treatment group}
\item{m1i}{Mean in treatment group}
\item{sd1i}{Standard deviation in treatment group}
\item{n2i}{Sample size in control group}
\item{m2i}{Mean in control group}
\item{sd2i}{Standard deviation in control group}
}
}
\usage{
data(dat_normand1999)
}
\description{
Data from a meta-analysis of continuous outcomes, including means and standard deviations
for treatment and control groups across different studies.
}
\details{
This dataset is suitable for meta-analysis of mean differences between two groups (treatment vs control).
}
\examples{
data(dat_normand1999)
head(dat_normand1999)

}
\keyword{datasets}
