% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_plot.R
\name{doi_plot}
\alias{doi_plot}
\title{DOI Plot for Meta-Analyses with Fewer Than 10 Studies}
\usage{
doi_plot(object, ...)
}
\arguments{
\item{object}{A meta-analysis object of class `meta_ratio`, `meta_mean`, or `meta_prop`.}

\item{...}{Additional arguments passed to `metasens::doiplot()`.}
}
\value{
Invisible NULL. Displays a plot to the active graphics device.
}
\description{
Displays a DOI plot and LFK index using metasens::doiplot(). Recommended for meta-analyses with <10 studies.
}
