% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fleiss1993bin.R
\docType{data}
\name{Fleiss1993bin}
\alias{Fleiss1993bin}
\title{Fleiss1993bin dataset}
\format{
A data frame with binary event outcomes.
}
\source{
Adapted from the `meta` package.
}
\usage{
Fleiss1993bin
}
\description{
Aspirin after myocardial infarction — binary data.
}
\keyword{datasets}
