% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_propulmeta.R
\docType{data}
\name{Olkin95}
\alias{Olkin95}
\title{Example dataset: Olkin95}
\format{
A data frame with 70 rows and 6 variables:
\describe{
\item{author}{Author name}
\item{year}{Year of publication}
\item{event.e}{Number of events in the experimental group}
\item{n.e}{Total participants in the experimental group}
\item{event.c}{Number of events in the control group}
\item{n.c}{Total participants in the control group}
}
}
\source{
Extracted from the \code{meta} package for demonstration purposes.
}
\usage{
data(Olkin95)
}
\description{
This dataset contains event counts from a set of studies comparing treatment and control groups.
}
\keyword{datasets}
