% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_propulmeta.R
\docType{data}
\name{Pagliaro1992}
\alias{Pagliaro1992}
\title{Portal Vein Thrombosis After Splenectomy Dataset}
\format{
An object of class \code{data.frame} (inherits from \code{pairwise}) with 28 rows and 8 columns.
}
\source{
Pagliaro et al., 1992. Adapted for meta-analysis teaching purposes.
}
\usage{
data(Pagliaro1992)
}
\description{
This dataset contains data from studies evaluating the incidence of portal vein thrombosis
following splenectomy.
}
\details{
#' @format A data frame with 28 rows and 8 variables:
\describe{
\item{id}{Study identifier}
\item{treat.exp}{Experimental treatment group label}
\item{logOR}{Log odds ratio}
\item{selogOR}{Standard error of the log odds ratio}
\item{bleed.exp}{Number of bleeding events in the experimental group}
\item{n.exp}{Sample size in the experimental group}
\item{bleed.plac}{Number of bleeding events in the placebo/control group}
\item{n.plac}{Sample size in the placebo/control group}
}

Useful for demonstrating meta-analyses involving rare event binary outcomes.
}
\examples{
data(Pagliaro1992)
head(Pagliaro1992)
}
\keyword{datasets}
