% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_meta.R
\name{plot_meta}
\alias{plot_meta}
\title{Generic forest plot wrapper for all meta-analysis types}
\usage{
plot_meta(x, ...)
}
\arguments{
\item{x}{A `meta_ratio`, `meta_mean`, or `meta_prop` object}

\item{...}{Additional arguments passed to the relevant plot function}
}
\value{
A forest plot (displayed or saved to file)
}
\description{
Dispatches to the appropriate plotting function based on class
}
