% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_meta_prop.R
\name{plot_meta_prop}
\alias{plot_meta_prop}
\title{Forest plot for meta-analysis of proportions (publication-ready)}
\usage{
plot_meta_prop(
  x,
  filename = NULL,
  save_as = c("viewer", "pdf", "png"),
  layout = "revman5",
  height = NULL,
  width = NULL,
  ...
)
}
\arguments{
\item{x}{A `meta_prop` object from `propulmeta::meta_prop()`}

\item{filename}{Output file name (used only if `save_as` is "pdf" or "png")}

\item{save_as}{Either "pdf", "png", or "viewer" (default = "viewer")}

\item{layout}{Forest plot layout: "revman5", "default", "JAMA", etc.}

\item{...}{Additional arguments passed to `meta::forest()`}
}
\value{
A forest plot either displayed in viewer or saved to file
}
\description{
Forest plot for meta-analysis of proportions (publication-ready)
}
\keyword{internal}
