% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlodipine.R
\docType{data}
\name{amlodipine}
\alias{amlodipine}
\title{Example dataset: amlodipine}
\format{
A data frame with X rows and Y variables:
\describe{
  \item{author}{Author name}
  \item{year}{Year of publication}
  \item{event.e}{Number of events in experimental group}
  \item{n.e}{Total participants in experimental group}
  \item{event.c}{Number of events in control group}
  \item{n.c}{Total participants in control group}
}
}
\source{
Derived from clinical trial data sets used for meta-analysis examples.
}
\usage{
data(amlodipine)
}
\description{
This dataset contains information on outcomes related to amlodipine studies.
}
\keyword{datasets}
