% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{bubble_plot}
\alias{bubble_plot}
\title{Bubble Plot for Meta-Regression}
\usage{
bubble_plot(meta_reg_object, moderator = NULL, plot_all_levels = TRUE, ...)
}
\arguments{
\item{meta_reg_object}{A `meta_reg` object created with `meta_reg()`.}

\item{moderator}{Character. Name of moderator variable (required if multiple predictors).}

\item{plot_all_levels}{Logical. For categorical moderators, plot each dummy variable (default = TRUE).}

\item{...}{Additional graphical arguments passed to `metafor::regplot()`.}
}
\value{
A bubble plot to the graphics device.
}
\description{
Displays a bubble plot for a meta_reg object using metafor::regplot().
}
