% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_propulmeta.R
\docType{data}
\name{Fleiss1993bin}
\alias{Fleiss1993bin}
\title{Fleiss1993bin: Aspirin After Myocardial Infarction Dataset (Binary Outcomes)}
\format{
A data frame with 7 rows and 6 variables:
\describe{
\item{study}{Study label}
\item{year}{Publication year}
\item{d.asp}{Number of deaths in the aspirin (treatment) group}
\item{n.asp}{Number of participants in the aspirin (treatment) group}
\item{d.plac}{Number of deaths in the placebo (control) group}
\item{n.plac}{Number of participants in the placebo (control) group}
}
}
\source{
Fleiss, J.L. (1993). \emph{The statistical basis of meta-analysis}. Statistical Methods in Medical Research.
Adapted for use in the \code{meta} package.
}
\usage{
data(Fleiss1993bin)
}
\description{
This dataset contains binary outcome data from randomized controlled trials
investigating the effect of aspirin versus placebo on mortality after myocardial infarction.
}
\details{
This dataset is often used in examples to illustrate meta-analysis methods for binary outcomes,
such as risk ratios, odds ratios, and risk differences.
}
\examples{
data(Fleiss1993bin)
head(Fleiss1993bin)
}
\keyword{datasets}
