% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heterogeneity.R
\name{plot_heterogeneity}
\alias{plot_heterogeneity}
\title{Heterogeneity Plot (I2 or Tau2) from Leave-One-Out Meta-Analysis}
\usage{
plot_heterogeneity(
  object,
  stat = c("I2", "tau2"),
  save_as = c("viewer", "pdf", "png"),
  filename = NULL,
  width = 10,
  height = NULL,
  ...
)
}
\arguments{
\item{object}{A \code{meta_prop}, \code{meta_ratio}, or \code{meta_mean} object.}

\item{stat}{"I2" or "tau2" (default "I2").}

\item{save_as}{"viewer", "pdf", or "png".}

\item{filename}{Optional file name if saving.}

\item{width, height}{Dimensions for export (inches).}

\item{...}{Additional arguments for plot().}
}
\value{
A heterogeneity plot.
}
\description{
Heterogeneity Plot (I2 or Tau2) from Leave-One-Out Meta-Analysis
}
