% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_bias.R
\name{publication_bias}
\alias{publication_bias}
\title{Publication Bias Inspection for Meta-Analysis (Single Plot Only)}
\usage{
publication_bias(object, plot_method = NULL)
}
\arguments{
\item{object}{A meta-analysis object from meta_ratio, meta_mean, or meta_prop.}

\item{plot_method}{Character string: "original", "trimfill", "contour", or "limitmeta".}
}
\value{
A list with Egger and Begg test results and optional models.
}
\description{
Performs Egger’s and Begg’s tests, and optionally displays one publication bias plot.
}
