% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_influence.R
\name{table_influence}
\alias{table_influence}
\title{Table: Leave-One-Out Influence Analysis}
\usage{
table_influence(object, include_heterogeneity = TRUE)
}
\arguments{
\item{object}{A meta_prop, meta_ratio, or meta_mean object}

\item{...}{Additional arguments passed to gtsummary}
}
\value{
A gtsummary publication-ready table
}
\description{
Table: Leave-One-Out Influence Analysis
}
