% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_meta.R
\name{table_meta}
\alias{table_meta}
\title{Generate Meta-Analysis Summary Table}
\usage{
table_meta(meta_result)
}
\arguments{
\item{meta_result}{A meta-analysis object of class `meta_ratio`, `meta_mean`, or `meta_prop`.}
}
\value{
A `gt` table summarizing study-level and pooled results.
}
\description{
Dispatches to the appropriate internal function to generate a summary table
depending on the class of the meta-analysis object.
}
\examples{
table_meta(meta_result)  # meta_result can be of class meta_ratio, meta_mean, or meta_prop
}
